requirements.txt
aiogram==3.4.1
yt-dlp>=2025.1.0
python-dotenv==1.0.1
requests>=2.32.0
